core/pyproject.toml
[tool.poetry]
name = "test-nginx-parser"
version = "0.1.0"
description = ""
authors = ["Arman <armanghandilyan977@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1"
djangorestframework = "^3.15.2"
requests = "^2.32.3"
django-filter = "^24.3"
psycopg2 = "^2.9.9"
drf-yasg = "^1.21.7"
tqdm = "^4.66.5"
rich = "^13.7.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
